\name{multinomial_logistic_EM}
\alias{multinomial_logistic_EM}
\title{
Part of the EM algorithm for multinomial logit mixture 
}
\description{
Part of the EM algorithm for multinomial logit mixture 
}
\usage{
multinomial_logistic_EM(y, x, K, w_start, b_start, 
	maxIter = 1000, emthreshold = 1e-08, maxNR = 5, 
	nCores = NULL, verbose = FALSE, R0, method)
}
\arguments{
  \item{y}{
y
}
  \item{x}{
X
}
  \item{K}{
K
}
  \item{w_start}{
w
}
  \item{b_start}{
b
}
  \item{maxIter}{
max
}
  \item{emthreshold}{
em
}
  \item{maxNR}{
maxnr
}
  \item{nCores}{
nc
}
  \item{verbose}{
verb
}
  \item{R0}{
or
}
  \item{method}{
method
}
}
\value{
value
}
\author{
Panagiotis Papastamoulis
}


