\name{multinomialLogitMix-package}
\alias{multinomialLogitMix-package}
\docType{package}
\title{
\packageTitle{multinomialLogitMix}
}
\description{
\packageDescription{multinomialLogitMix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{multinomialLogitMix}
\packageIndices{multinomialLogitMix}
See the main function of the package: \code{\link{multinomialLogitMix}}, which wraps automatically calls to the MCMC sampler \code{\link{gibbs_mala_sampler_ppt}} and the EM algorithm \code{\link{mix_mnm_logistic}}. 
}
\author{
\packageAuthor{multinomialLogitMix}

Maintainer: \packageMaintainer{multinomialLogitMix}
}
\references{
Papastamoulis, P. (2022). Model-based clustering of replicated multinomial data. arXiv:2207.13984 [stat.ME]

Papastamoulis, P. and Iliopoulos, G. (2010). An Artificial Allocations Based Solution to the Label Switching Problem in Bayesian Analysis of Mixtures of Distributions. Journal of Computational and Graphical Statistics, 19(2), 313-331. http://www.jstor.org/stable/25703571

Papastamoulis, P. (2016). label.switching: An R Package for Dealing with the Label Switching Problem in MCMC Outputs. Journal of Statistical Software, Code Snippets, 69(1), 1-24. https://doi.org/10.18637/jss.v069.c01

Rousseau, J. and Mengersen, K. (2011), Asymptotic behaviour of the posterior distribution in overfitted mixture models. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 73: 689-710. https://doi.org/10.1111/j.1467-9868.2011.00781.x
}
\keyword{ package }
\seealso{
\code{\link{multinomialLogitMix}}, \code{\link{gibbs_mala_sampler_ppt}},\code{\link{mix_mnm_logistic}}
}

